web: filter Sentry errors reported not from the ErrorBoundary too
