+++ client/web/src/components/ErrorBoundary.test.tsx
@@ -1,11 +1,5 @@
-import * as sentry from '@sentry/browser'
-import { render } from '@testing-library/react'
 import React from 'react'
 import renderer from 'react-test-renderer'
-import sinon from 'sinon'
-
-import { AbortError } from '@sourcegraph/shared/src/api/util'
-import { HTTPStatusError } from '@sourcegraph/shared/src/backend/fetch'
 
 import { ErrorBoundary } from './ErrorBoundary'
 
@@ -21,19 +15,6 @@ const ThrowChunkError: React.FunctionComponent = () => {
     const ChunkError = new Error('Loading chunk 123 failed.')
     ChunkError.name = 'ChunkLoadError'
     throw ChunkError
-}
-
-const ThrowAbortError: React.FunctionComponent = () => {
-    throw new AbortError()
-}
-
-const ThrowNotAuthenticatedError: React.FunctionComponent = () => {
-    throw new Error('not authenticated')
-}
-
-const ThrowHTTPStatusError: React.FunctionComponent<{ status?: number }> = ({ status = 500 }) => {
-    const errorResponse = new Response(null, { status })
-    throw new HTTPStatusError(errorResponse)
 }
 
 describe('ErrorBoundary', () => {
@@ -69,70 +50,4 @@ describe('ErrorBoundary', () => {
                 )
                 .toJSON()
         ).toMatchSnapshot())
-
-    test('Sentry should capture HttpStatusError except for Server response errors (5xx)', () => {
-        const sentryCaptureException = sinon.stub(sentry, 'captureException')
-
-        render(
-            <ErrorBoundary location={null}>
-                <ThrowHTTPStatusError status={500} />
-            </ErrorBoundary>
-        )
-
-        sinon.assert.notCalled(sentryCaptureException)
-        sentryCaptureException.reset()
-
-        render(
-            <ErrorBoundary location={null}>
-                <ThrowHTTPStatusError status={400} />
-            </ErrorBoundary>
-        )
-
-        sinon.assert.calledOnce(sentryCaptureException)
-
-        const [capturedError] = sentryCaptureException.getCall(0).args
-        expect(capturedError).toBeInstanceOf(HTTPStatusError)
-        expect(capturedError).toHaveProperty('status', 400)
-
-        sentryCaptureException.restore()
-    })
-
-    test('Sentry should not capture AbortError', () => {
-        const sentryCaptureException = sinon.stub(sentry, 'captureException')
-
-        render(
-            <ErrorBoundary location={null}>
-                <ThrowAbortError />
-            </ErrorBoundary>
-        )
-
-        sinon.assert.notCalled(sentryCaptureException)
-        sentryCaptureException.restore()
-    })
-
-    test('Sentry should not capture ChunkLoadError', () => {
-        const sentryCaptureException = sinon.stub(sentry, 'captureException')
-
-        render(
-            <ErrorBoundary location={null}>
-                <ThrowChunkError />
-            </ErrorBoundary>
-        )
-
-        sinon.assert.notCalled(sentryCaptureException)
-        sentryCaptureException.restore()
-    })
-
-    test('Sentry should not capture not authenticated error', () => {
-        const sentryCaptureException = sinon.stub(sentry, 'captureException')
-
-        render(
-            <ErrorBoundary location={null}>
-                <ThrowNotAuthenticatedError />
-            </ErrorBoundary>
-        )
-
-        sinon.assert.notCalled(sentryCaptureException)
-        sentryCaptureException.restore()
-    })
 })
