+++ client/web/src/enterprise/main.tsx
@@ -5,7 +5,7 @@
 
 import '@sourcegraph/shared/src/polyfills'
 
-import '../sentry'
+import '../sentry/init'
 
 import React from 'react'
 import { render } from 'react-dom'
